+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## unreleased
+
+- Added a backup config file, where the device seed is encrypted with the user provided passphrase. This is downloaded in the same step as the main config file. [#140](https://github.com/Holo-Host/hpos-config/pull/140)
+
 ## 0.1.0
 
 - Updated Config::V2 to include device_seed and remove the `encrypted_key` option.